Working database implementation, comments aren't accurate

+++ Backend/app/src/main/java/resources/ResourceManager.java
@@ -2,6 +2,7 @@ package resources;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import jakarta.persistence.*;
+import org.springframework.beans.factory.annotation.Autowired;
 import troops.TroopManagerSerializer;
 
 import java.util.ArrayList;
@@ -16,13 +17,15 @@ public class ResourceManager {
     //Usefulness of child resource classes isn't currently apparent but they will eventaully have other properties
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long playerID;
+    private Integer playerID;
 
     @OneToMany(mappedBy = "resourceManager", cascade = CascadeType.ALL)
     private List<Resource> resourceManager;
 
+
+
     // Every player starts with quantity of resource described below, allows introduction to building/recruiting/research system
-    public ResourceManager(long playerId) {
+    public ResourceManager(Integer playerId) {
         this.playerID = playerId;
         this.resourceManager = new ArrayList<>();
         initializeResources();
